Initial commit for the duck player client
I hope this will turn out fine in the end.
I've now got basic code layout and some of the basic controls
working in the UI.
The idea is to have an extensible "player", which will
not have its own backend, but it would still be possible
to add one later.