OK, Synchronization looks OK now, But, is it?
Resorted to a pretty simple scheme. Seems to work and I've been thinking
about corner cases -- none come to mind right now.
This will need to be better analyzed later.
I've put a lot of debugging code in there. Will remove it later, but it
won't hurt to keep it for now.